pyproject.toml
[project]
name = "dca"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Håkon Åmdal", email = "hakon@aamdal.com" }
]
dependencies = [
    "pandas>=2.2.3",
    "requests>=2.32.3",
    "colorama>=0.4.6",
    "pyyaml>=6.0.2",
    "tabulate>=0.9.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/dca"]
